refactor/142: 단위부터 통합테스트까지 그 과정에서 리팩토링

+++ src/main/java/com/homesweet/homesweetback/domain/order/dto/request/CreateOrderRequest.java
@@ -18,6 +18,6 @@ public record CreateOrderRequest(
             Long skuId,
 
             @Min(value = 1, message = "주문 수량은 1개 이상이어야 합니다.")
-            int quantity
+            long quantity
     ){}
 }
